One step closer the Production ready CIA.
The if the first client relodads the page while it is generating,
There will be no connection available from the server to the client,
when the server is about to tranfer the results trough the wire.
Because of this at the ob_callback the $output varibale will be empty.
So the client is forced to reload the page.